Face-Recognition-with-OpenCV-Python-DL/requirements.txt
# Các thư viện cần thiết cho dự án nhận diện khuôn mặt
flask==2.0.1
flask-socketio==5.1.1
opencv-python==4.5.3.56
numpy==1.21.2
deepface==0.0.75
Pillow==8.3.2
python-engineio==4.2.1
python-socketio==5.4.0